TMBF: Widen ADC:OFFSET
Strictly speaking these are 17 bit numbers, but of course we only want
this many bits when testing.

git-svn-id: svn+ssh://serv2/home/subversion/repos/controls/diamond/trunk/ioc/TMBF@124097 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tmbfApp/src/adc_dac.c
@@ -181,7 +181,7 @@ bool initialise_adc_dac(void)
     publish_minmax("DAC", &dac_min_max, 1 << 13);
 
     /* Offset control for ADC. */
-    PUBLISH_WF_ACTION_P(short, "ADC:OFFSET", 4, hw_write_adc_offsets);
+    PUBLISH_WF_ACTION_P(int, "ADC:OFFSET", 4, hw_write_adc_offsets);
     PUBLISH_WAVEFORM(float, "ADC:FILTER", 12, write_filter_coeffs,
         .context = &adc_min_max, .persist = true);
     PUBLISH_WRITER_P(mbbo, "ADC:FILTER:DELAY", hw_write_adc_filter_delay);
